langgraph_groq_project/pyproject.toml
[project]
name = "langgraph_groq_project"
version = "0.1.0"
description = "LangGraph + LangChain + Groq project for SQL data analysis"
requires-python = ">=3.10"
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-groq",
    "langgraph",
    "python-dotenv",
    "pandas",
    "pandasql",
    "typing-extensions",
    "pyyaml"
]
